pyproject.toml
[project]
name = "court-jester-selector-bot"
version = "0.1.0"
description = "🃏 A fun Telegram bot that crowns a random friend as the day's entertainer based on weighted randomization. All hail the royal jester! 🎪"
authors = [
    { name = "j-about", email = "142051449+j-about@users.noreply.github.com" },
]
requires-python = ">=3.13.13"
dependencies = [
    "python-telegram-bot==22.7",
    "SQLModel==0.0.38",
    "asyncpg==0.31.0",
    "psycopg[binary]==3.3.3",
    "alembic==1.18.4",
    "pydantic-settings==2.13.1",
    "sentry-sdk==2.58.0",
]

[project.scripts]
court-jester-bot = "main:main"

[dependency-groups]
dev = [
    "pytest==9.0.3",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.1.0",
    "ruff==0.15.10",
    "ty==0.0.31",
]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]

[tool.ty.src]
include = [
    "config.py",
    "main.py",
    "models.py",
    "database.py",
    "queries.py",
    "handlers",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov --cov-report=term-missing --cov-fail-under=85"

[tool.coverage.run]
source = ["config", "models", "database", "queries", "observability", "handlers", "utils"]
omit = ["tests/*", "alembic/*", "main.py"]
